% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsSim-package.R
\docType{package}
\name{amsSim-package}
\alias{amsSim}
\alias{amsSim-package}
\title{amsSim: Adaptive Multilevel Splitting e Monte Carlo}
\description{
Simulation and pricing routines for rare-event options using Adaptive Multilevel Splitting and standard Monte Carlo under Black–Scholes and Heston models. Core routines are implemented in C++ via RcppArmadillo with R wrappers.
}
\seealso{
Useful links:
\itemize{
  \item \url{Mettere la mia}
  \item Report bugs at \url{Mettere la mia}
}

}
\author{
\strong{Maintainer}: Riccardo Gozzo \email{gozzoriccardo0@gmail.com}

}
\keyword{internal}
